requirements.txt
Flask==2.0.1
Django==3.2.5
requests==2.25.1
numpy==1.21.0
pandas==1.3.0
djangorestframework==3.12.4
gunicorn==20.1.0
